// src/enums/expertise.enum.ts
export enum ExpertiseEnum {
  WOMEN_AND_CHILD_BIRTH = "زنان، زایمان و نازایی",
  OPHTHALMOLOGY = "چشم پزشکی",
  GASTROENTEROLOGY = "گوارش و معده",
  KIDNEY_AND_URINARY_TRACT = "کلیه و مجاری ادراری",
  SPEECH_THERAPY = "گفتار درمانی",
  CHILDREN_DISEASES = "بیماری‌های کودکان",
  PUBLIC_AUCTION = "جراحی عمومی",
  NEONATAL_MEDICINE = "طب نوزادی",
  ORTHOPEDIC_SPECIALIST = "متخصص ارتوپدی",
  CARDIOVASCULAR_DISEASES = "بیماری‌های قلب و عروق",
}